@@ -35,33 +35,36 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand"
   },
   "dependencies": {
-    "@nestjs/common": "^6.0.0",
-    "@nestjs/core": "^6.0.0",
-    "reflect-metadata": "^0.1.12",
-    "rimraf": "^2.6.2",
-    "rxjs": "^6.3.3",
-    "ssh2-sftp-client": "^2.5.0"
+    "@nestjs/common": "^7.0.7",
+    "@nestjs/core": "^7.0.7",
+    "reflect-metadata": "^0.1.13",
+    "rimraf": "^3.0.2",
+    "rxjs": "^6.5.4",
+    "ssh2-sftp-client": "^5.1.2"
   },
   "devDependencies": {
-    "@nestjs/testing": "6.1.1",
-    "@semantic-release/changelog": "^3.0.4",
-    "@semantic-release/git": "^7.0.16",
-    "@types/jest": "24.0.11",
-    "@types/node": "^11.13.4",
-    "@types/ssh2-sftp-client": "^2.5.1",
-    "@types/supertest": "2.0.7",
-    "husky": "^3.0.3",
-    "jest": "24.7.1",
-    "lint-staged": "^9.2.1",
-    "prettier": "1.17.0",
-    "semantic-release": "^15.13.21",
+    "@nestjs/testing": "7.0.7",
+    "@semantic-release/changelog": "^5.0.1",
+    "@semantic-release/git": "^9.0.0",
+    "@types/jest": "25.1.4",
+    "@types/node": "^12.7.5",
+    "@types/ssh2-sftp-client": "^4.1.3",
+    "@types/supertest": "2.0.8",
+    "import-sort-style-eslint": "^6.0.0",
+    "husky": "^4.2.3",
+    "jest": "25.2.4",
+    "lint-staged": "^10.1.1",
+    "prettier": "2.0.2",
+    "prettier-plugin-import-sort": "0.0.4",
+    "semantic-release": "^17.0.4",
     "supertest": "4.0.2",
-    "ts-jest": "24.0.2",
-    "ts-node": "8.1.0",
-    "tsc-watch": "2.2.1",
-    "tsconfig-paths": "3.8.0",
-    "tslint": "5.16.0",
-    "typescript": "3.4.3"
+    "ts-jest": "25.3.0",
+    "ts-node": "8.8.1",
+    "tsc-watch": "4.2.3",
+    "tsconfig-paths": "3.9.0",
+    "tslint": "6.1.0",
+    "tslint-config-prettier": "^1.18.0",
+    "typescript": "3.8.3"
   },
   "release": {
     "branch": "master",
@@ -144,6 +147,11 @@
       "prettier --write \"src/**/*.ts\"",
       "git add"
     ]
+  },
+  "importSort": {
+    ".js, .jsx, .ts, .tsx": {
+      "parser": "typescript"
+    }
   },
   "jest": {
     "moduleFileExtensions": [
